==> code/data_collection/5 - Twitter Scraping.ipynb <==
This notebook contains code used to configure and consolidate data from the Twitter API using urls found in TheFlipSide articles. In contrast to other notebooks, this was run in two parts: sections preceding "Consolidating Tweets", and then the final section after fetching the data using the pull_twitter tool.

# Imports 

import pandas as pd
import numpy as np

import json
import yaml
import os
from datetime import datetime

import requests
import urlexpander

from ast import literal_eval

from tqdm import tqdm
tqdm.pandas()

import time

import urlexpander

# Settings and Configuration 

SRC_DATA   = '../../data/webpage_data/full_flipside_data_clean.csv'
TWEET_DATA = '../../data/twitter_data/url_query_list'
OUTPUT_DIR = '../../data/twitter_data/search/'
FINAL_OUTPUT_FP = '../../data/twitter_data/all_tweet_data.csv'

CLIP_URLS = 8 # Take first N urls to attempt to expand for scraping Twitter
MAX_URLS  = 5 # Maximum number of urls to use in scraping Twitter

# Data Loading 

Load and clean the original article data 

art_data = pd.read_csv(SRC_DATA)

art_data['linked_arts_clean'] = art_data['linked_arts_clean'].str.replace('\n', ',')

art_data['linked_arts_clean'] = art_data['linked_arts_clean'].apply(literal_eval)

art_data['num_news'] = art_data['linked_arts_clean'].apply(len)

# Creating Query List

tweet_queries = [url.split('?')[0] for url_list in art_data['linked_arts_clean'].values for url in url_list]

tweet_queries = ['url:\"' + q + '\" -is:reply -is:retweet -is:verified' for q in tweet_queries]

with open(TWEET_DATA, 'w') as f:
    f.write('\n'.join(tweet_queries))

# Run Scraping

The twitter scraping is handled by the `scrape_twitter.sh` bash script in the same directory as this notebook. Run the following command to generate the scraped files.

```
    ./scrape_twitter.sh  "../../data/twitter_data/url_query_list"
```

# Consolidating Tweets 

Align the queries and Twitter API Results to map tweets back to the database

# Create map of urls to titles
full_queries = {url.split('?')[0]: title + '|' + date for (title, date), url_list in zip(zip(art_data['title'], art_data['date']), art_data['linked_arts_clean']) for url in url_list}

# Retrieve the ordered list of tweet urls
with open(TWEET_DATA, 'r') as f:
    queries = f.readlines()

# Generate unsorted list of data directories
data_dirs = [OUTPUT_DIR + resdir for resdir in os.listdir(OUTPUT_DIR) if resdir[-4:] != '.zip']

# Sort the Twitter API results using the datetime in the directory name
data_dirs.sort(key = lambda fn: datetime.strptime(fn.split('/')[-1], '%Y-%m-%d %H.%M.%S'))

data_fps = [data_dir + '/data_tweets.csv' for data_dir in data_dirs]
query_urls = [query.split(' ')[0][5:-1] for query in queries]

url_to_data = {query_url: data_fp for query_url, data_fp in zip(query_urls, data_fps)}

# Ensure all urls are present in both mapping dictionaries
assert set(url_to_data.keys()) == set(full_queries.keys())

data_to_title_date = {url_to_data[url]: title_date for url, title_date in full_queries.items() }

# Extract the title and date columns separately and prepare to consolidate data
all_tweets = []
for data_fp, title_date_q in data_to_title_date.items():
    try:
        data = pd.read_csv(data_fp)
        data['title_q'] = title_date_q.split('|')[0]
        data['date_q']  = title_date_q.split('|')[-1]
        all_tweets.append(data)
    except:
        print(f'{title_date_q.split("|")[0]} had no results for a url')

# Consolidate and save tweet data, filtering by english tweets
all_tweets_df = pd.concat(all_tweets, axis = 0)
all_tweets_df = all_tweets_df[all_tweets_df['lang'] == 'en']
all_tweets_df.to_csv(FINAL_OUTPUT_FP, index = None)
